long

+++ MogulMoves/app/src/main/java/com/example/mogulmoves/ExperimentSerializer.java
@@ -91,15 +91,15 @@ public class ExperimentSerializer implements Serializer<Experiment> {
         experiment.setActive(active);
 
         try {
-            for(int trial: trials){
-                experiment.addTrial(trial);
+            for(long trial: trials){
+                experiment.addTrial((int) trial);
             }
         } catch (java.lang.NullPointerException e) {
         }
 
         try {
-            for (int message : messages) {
-                experiment.addMessage(message);
+            for (long message: messages) {
+                experiment.addMessage((int) message);
             }
         } catch (java.lang.NullPointerException e) {
         }
